+++ lenstools/noise.py
@@ -16,14 +16,9 @@ from convergence import ConvergenceMap
 
 import numpy as np
 
-#FFT engines
-from numpy.fft import fftfreq,irfft2
-
-#Check if rfftfreq is implemented (requires numpy>=1.8)
-try:
-	from numpy.fft import rfftfreq
-except ImportError:
-	from utils import rfftfreq
+#FFT engine
+from .fft import NUMPYFFTPack
+fftengine = NUMPYFFTPack()
 
 from scipy import interpolate
 
@@ -109,8 +104,8 @@ class GaussianNoiseGenerator(object):
 
 		#Assert the shape of the blueprint, to tune the right size for the fourier transform
 		lpix = 360.0/self.side_angle.to(deg).value
-		lx = rfftfreq(self.shape[0]) * self.shape[0] * lpix
-		ly = fftfreq(self.shape[0]) * self.shape[0] * lpix
+		lx = fftengine.rfftfreq(self.shape[0]) * self.shape[0] * lpix
+		ly = fftengine.fftfreq(self.shape[0]) * self.shape[0] * lpix
 
 		#Compute the multipole moment of each FFT pixel
 		l = np.sqrt(lx[np.newaxis,:]**2 + ly[:,np.newaxis]**2)
@@ -168,7 +163,7 @@ class GaussianNoiseGenerator(object):
 
 		#Generate a random Fourier realization and invert it
 		ft_map = self._fourierMap(power_func,**kwargs)
-		noise_map = irfft2(ft_map)
+		noise_map = fftengine.irfft2(ft_map)
 
 		return ConvergenceMap(noise_map,self.side_angle)